rename ICE watcher script; add script to run jobs from queue.
git-svn-id: svn+ssh://astrometry.net/svn/trunk/src/astrometry@10283 03a3cea6-2c03-0410-ac9b-9271b1c66c29

+++ net/portal/job.py
@@ -751,9 +751,15 @@ class Job(models.Model):
     def submit_job_or_submission(j):
         os.umask(07)
         j.create_job_dir()
-        # enqueue by creating a symlink in the job queue directory.
-        jobdir = j.get_job_dir()
-        link = settings.JOB_QUEUE_DIR + j.get_id()
-        if os.path.exists(link):
-            os.unlink(link)
-        os.symlink(jobdir, link)
+
+        from astrometry.net.portal.queue import *
+        QueuedJob.submit_job_or_submission(j)
+
+        # watcher-based
+        if False:
+            # enqueue by creating a symlink in the job queue directory.
+            jobdir = j.get_job_dir()
+            link = settings.JOB_QUEUE_DIR + j.get_id()
+            if os.path.exists(link):
+                os.unlink(link)
+            os.symlink(jobdir, link)
